+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Upcoming Version
 
+## 4.6.6 "All development stems from open source"
+
+###### [Server](https://github.com/floere/picky/tree/master/server)
+
+- Experimental update: Stemming.
+- New tokenizer (indexing/searching) option: `stems\_with`. Give it a thing that responds to `#stem(text) # => stemmed\_text`.
+- See [https://github.com/floere/picky/blob/master/server/spec/functional/stemming_spec.rb](https://github.com/floere/picky/blob/master/server/spec/functional/stemming_spec.rb) for examples.
+
 ## 4.6.5 "Help is on the way"
 
 ###### [Server](https://github.com/floere/picky/tree/master/server)